#124 [refactor] 바코드검색, 텍스트검색 유즈케이스 적용
